+++ src/index.ts
@@ -15,6 +15,7 @@ export { Scenario } from './Scenario';
 export { Messenger } from './Messenger/Adapter';
 export { Telegram } from './Telegram/Adapter';
 export { Response as IResponse } from './Response';
-export { MessageMiddleware } from './Middlewares/Message';
+export { MessageMiddleware, ISubscriber, ISubscriberCallback } from './Middlewares/Message';
+export { SubscriberContext } from './SubscriberContext';
 export { IEvent } from './Events/Event';
 export * from './Events/SceneEvents';